+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Released schemas are now shipped with the package, so an internet connection
   is no longer necessary for local validation. 
+* Added `subset_task_id_names()` function to subset task ID names from a character vector of column names (#149).
+* Added functions `subset_task_id_cols()` and `subset_std_cols()` to subset a `model_out_tbl` or submission `tbl` to task ID or standard (non-task ID) columns respectively (#149).
 
 # hubUtils 0.3.0